requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
pdf2image==1.17.0
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
pandas==2.2.1
beautifulsoup4==4.12.3
loguru==0.7.2
aiohttp==3.9.3
pydantic==2.6.1
python-jose==3.3.0
httpx==0.26.0
python-docx==1.1.0
mammoth==1.6.0